pyproject.toml
[project]
name = "rag"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
    "chromadb>=1.2.1",
    "gradio>=5.49.1",
    "ipykernel>=7.0.1",
    "ipywidgets>=8.1.7",
    "langchain>=1.0.2",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4",
    "langchain-core>=1.0.0",
    "langchain-huggingface>=1.0.0",
    "langchain-openai>=1.0.1",
    "langchain-text-splitters>=1.0.0",
    "litellm>=1.78.6",
    "matplotlib>=3.10.7",
    "nbformat>=5.10.4",
    "openai>=2.6.0",
    "openai-agents>=0.4.1",
    "plotly>=6.3.1",
    "python-dotenv>=1.1.1",
    "ratelimit>=2.2.1",
    "sentence-transformers>=5.1.2; sys_platform != 'darwin' or platform_machine != 'x86_64'",
    "sentence-transformers>=3.0,<5; sys_platform == 'darwin' and platform_machine == 'x86_64'",
    "numpy<2; sys_platform == 'darwin' and platform_machine == 'x86_64'",
    "tiktoken>=0.12.0",
    "torch>=2.8.0; sys_platform != 'darwin' or platform_machine != 'x86_64'",
    "torch>=2.2.0,<2.3; sys_platform == 'darwin' and platform_machine == 'x86_64'",
]

[tool.uv]
required-environments = [
    "sys_platform == 'linux'",
    "sys_platform == 'darwin' and platform_machine == 'arm64'",
    "sys_platform == 'darwin' and platform_machine == 'x86_64'",
    "sys_platform == 'win32'",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", marker = "sys_platform == 'darwin' and platform_machine == 'x86_64'" },
]

